patience=10 -> patience=3, adds check_early_stop, refactors train code
