+++ egs/ami/asr1/run.sh
@@ -281,7 +281,7 @@ if [[ ${stage} -le 3 && ${use_lm} == true ]]; then
         --unit ${lm_units} \
         --opt ${lm_opt} \
         --batchsize ${lm_batchsize} \
-        --epoch ${lm_epochs} \
+        --epochs ${lm_epochs} \
         --patience ${lm_patience} \
         --maxlen ${lm_maxlen} \
         --dict ${lmdict}
