+++ egs/csj/asr1/local/run_spm.sh
@@ -205,7 +205,7 @@ if [ ${stage} -le 3 ]; then
         --tensorboard-dir tensorboard/${lmexpname} \
         --train-label ${lmdatadir}/train.txt \
         --valid-label ${lmdatadir}/valid.txt \
-        --epoch 40 \
+        --epochs 40 \
         --batchsize 256 \
         --dict ${dict}
 fi
