+++ egs/fisher_swbd/asr1/run.sh
@@ -249,7 +249,7 @@ if [ ${stage} -le 3 ]; then
         --unit ${lm_units} \
         --opt ${lm_opt} \
         --batchsize ${lm_batchsize} \
-        --epoch ${lm_epochs} \
+        --epochs ${lm_epochs} \
         --patience ${lm_patience} \
         --maxlen ${lm_maxlen} \
         --dict ${dict}
