+++ espnet/bin/lm_train.py
@@ -53,11 +53,11 @@ def main(args):
                         help='Optimizer')
     parser.add_argument('--batchsize', '-b', type=int, default=300,
                         help='Number of examples in each mini-batch')
-    parser.add_argument('--epoch', '-e', type=int, default=20,
+    parser.add_argument('--epochs', '-e', type=int, default=20,
                         help='Number of sweeps over the dataset to train')
     parser.add_argument('--early-stop-criterion', default='validation/main/loss', type=str, nargs='?',
                         help="Value to monitor to trigger an early stopping of the training")
-    parser.add_argument('--patience', default=0, type=int, nargs='?',
+    parser.add_argument('--patience', default=3, type=int, nargs='?',
                         help="Number of epochs to wait without improvement before stopping the training")
     parser.add_argument('--gradclip', '-c', type=float, default=5,
                         help='Gradient norm threshold to clip')
